+++ src/main/java/net/yxiao233/ifeu/common/compact/jei/ModRecipeType.java
@@ -10,5 +10,6 @@ public class ModRecipeType {
     public static RecipeType<ArcaneDragonEggForgingRecipe> ARCANE_DRAGON_EGG_FORGING = RecipeType.create(nameSpace,"arcane_dragon_egg_forging", ArcaneDragonEggForgingRecipe.class);
     public static RecipeType<BlockRightClickRecipe> BLOCK_RIGHT_CLICK = RecipeType.create(nameSpace,"block_right_click", BlockRightClickRecipe.class);
     public static RecipeType<DragonStarGeneratorRecipe> DRAGON_STAR_GENERATOR = RecipeType.create(nameSpace,"dragon_star_generator", DragonStarGeneratorRecipe.class);
-    public static RecipeType<FluidCraftingTableRecipe> FLUID_CRAFTING_TABLE = RecipeType.create(nameSpace,"fluid_crafting_table", FluidCraftingTableRecipe.class);
+    public static RecipeType<ShapedRecipe> SHAPED = RecipeType.create(nameSpace,"shaped", ShapedRecipe.class);
+    public static RecipeType<ShapelessRecipe> SHAPELESS = RecipeType.create(nameSpace,"shapeless", ShapelessRecipe.class);
 }